# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Self-hosted media downloader API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DownloadStatus = typeof DownloadStatus[keyof typeof DownloadStatus];


export const DownloadStatus = {
  pending: 'pending',
  downloading: 'downloading',
  completed: 'completed',
  failed: 'failed',
} as const;

export type DownloadFormat = typeof DownloadFormat[keyof typeof DownloadFormat];


export const DownloadFormat = {
  mp4: 'mp4',
  mp3: 'mp3',
  webm: 'webm',
  best: 'best',
} as const;

export interface Download {
  id: number;
  url: string;
  /** @nullable */
  title: string | null;
  /** @nullable */
  thumbnail?: string | null;
  status: DownloadStatus;
  format: DownloadFormat;
  /** @nullable */
  quality?: string | null;
  /** @nullable */
  fileSize?: number | null;
  /** @nullable */
  filePath?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  duration?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export type DownloadInputFormat = typeof DownloadInputFormat[keyof typeof DownloadInputFormat];


export const DownloadInputFormat = {
  mp4: 'mp4',
  mp3: 'mp3',
  webm: 'webm',
  best: 'best',
} as const;

export interface DownloadInput {
  /** @minLength 1 */
  url: string;
  format: DownloadInputFormat;
  quality?: string;
}

export interface ClearCompletedResponse {
  deleted: number;
}

export interface Stats {
  total: number;
  pending: number;
  downloading: number;
  completed: number;
  failed: number;
  totalSizeBytes: number;
}

export type ScheduleFormat = typeof ScheduleFormat[keyof typeof ScheduleFormat];


export const ScheduleFormat = {
  mp4: 'mp4',
  mp3: 'mp3',
  webm: 'webm',
  best: 'best',
} as const;

export interface Schedule {
  id: number;
  label: string;
  url: string;
  format: ScheduleFormat;
  /** @nullable */
  quality?: string | null;
  cronExpr: string;
  enabled: boolean;
  /** @nullable */
  lastRunAt?: string | null;
  /** @nullable */
  nextRunAt?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export type ScheduleInputFormat = typeof ScheduleInputFormat[keyof typeof ScheduleInputFormat];


export const ScheduleInputFormat = {
  mp4: 'mp4',
  mp3: 'mp3',
  webm: 'webm',
  best: 'best',
} as const;

export interface ScheduleInput {
  /** @minLength 1 */
  label: string;
  url: string;
  format: ScheduleInputFormat;
  quality?: string;
  cronExpr: string;
  enabled?: boolean;
}

export type ScheduleUpdateFormat = typeof ScheduleUpdateFormat[keyof typeof ScheduleUpdateFormat];


export const ScheduleUpdateFormat = {
  mp4: 'mp4',
  mp3: 'mp3',
  webm: 'webm',
  best: 'best',
} as const;

export interface ScheduleUpdate {
  label?: string;
  url?: string;
  format?: ScheduleUpdateFormat;
  quality?: string;
  cronExpr?: string;
  enabled?: boolean;
}

export interface BatchDownloadInput {
  items: DownloadInput[];
}

export interface ScheduleRunResponse {
  dispatched: number;
  jobs: Download[];
}

export type ListDownloadsParams = {
status?: ListDownloadsStatus;
format?: string;
};

export type ListDownloadsStatus = typeof ListDownloadsStatus[keyof typeof ListDownloadsStatus];


export const ListDownloadsStatus = {
  pending: 'pending',
  downloading: 'downloading',
  completed: 'completed',
  failed: 'failed',
} as const;