# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Self-hosted media downloader API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: downloads
    description: Download job operations
  - name: queue
    description: Queue and stats operations
  - name: schedules
    description: Scheduled download operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /downloads:
    get:
      operationId: listDownloads
      tags: [downloads]
      summary: List all downloads
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, downloading, completed, failed]
        - name: format
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of download jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Download"
    post:
      operationId: createDownload
      tags: [downloads]
      summary: Create a new download job (auto-expands channel/playlist URLs)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DownloadInput"
      responses:
        "201":
          description: Download job(s) created
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Download"
                  - type: array
                    items:
                      $ref: "#/components/schemas/Download"
        "400":
          description: Invalid request
        "422":
          description: Could not expand channel/playlist

  /downloads/batch:
    post:
      operationId: batchCreateDownloads
      tags: [downloads]
      summary: Create multiple download jobs at once
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchDownloadInput"
      responses:
        "201":
          description: Download jobs created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Download"
        "400":
          description: Invalid request

  /downloads/{id}:
    get:
      operationId: getDownload
      tags: [downloads]
      summary: Get a download job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Download job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Download"
        "404":
          description: Not found
    delete:
      operationId: deleteDownload
      tags: [downloads]
      summary: Delete a download job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /downloads/completed:
    delete:
      operationId: clearCompleted
      tags: [downloads]
      summary: Delete all completed downloads and their files
      responses:
        "200":
          description: Number of deleted downloads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClearCompletedResponse"

  /downloads/{id}/retry:
    post:
      operationId: retryDownload
      tags: [downloads]
      summary: Retry a failed download
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Retried download job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Download"
        "404":
          description: Not found

  /downloads/{id}/file:
    get:
      operationId: downloadFile
      tags: [downloads]
      summary: Download the completed file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File stream
        "404":
          description: Not found or not completed

  /stats:
    get:
      operationId: getStats
      tags: [queue]
      summary: Get download queue statistics
      responses:
        "200":
          description: Queue stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

  /recent:
    get:
      operationId: getRecent
      tags: [queue]
      summary: Get recently completed downloads
      responses:
        "200":
          description: Recent downloads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Download"

  /schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: List all scheduled downloads
      responses:
        "200":
          description: List of schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Create a new scheduled download
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInput"
      responses:
        "201":
          description: Schedule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "400":
          description: Invalid request

  /schedules/{id}:
    get:
      operationId: getSchedule
      tags: [schedules]
      summary: Get a schedule by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "404":
          description: Not found
    put:
      operationId: updateSchedule
      tags: [schedules]
      summary: Update a schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleUpdate"
      responses:
        "200":
          description: Updated schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "404":
          description: Not found
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      summary: Delete a schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /schedules/{id}/run:
    post:
      operationId: runScheduleNow
      tags: [schedules]
      summary: Run a scheduled download immediately
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Jobs dispatched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleRunResponse"
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Download:
      type: object
      required:
        - id
        - url
        - title
        - status
        - format
        - createdAt
      properties:
        id:
          type: integer
        url:
          type: string
        title:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, downloading, completed, failed]
        format:
          type: string
          enum: [mp4, mp3, webm, best]
        quality:
          type: ["string", "null"]
        fileSize:
          type: ["integer", "null"]
        filePath:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        progress:
          type: ["number", "null"]
        duration:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]

    DownloadInput:
      type: object
      required:
        - url
        - format
      properties:
        url:
          type: string
          minLength: 1
        format:
          type: string
          enum: [mp4, mp3, webm, best]
        quality:
          type: string

    ClearCompletedResponse:
      type: object
      required:
        - deleted
      properties:
        deleted:
          type: integer

    Stats:
      type: object
      required:
        - total
        - pending
        - downloading
        - completed
        - failed
        - totalSizeBytes
      properties:
        total:
          type: integer
        pending:
          type: integer
        downloading:
          type: integer
        completed:
          type: integer
        failed:
          type: integer
        totalSizeBytes:
          type: integer

    Schedule:
      type: object
      required:
        - id
        - label
        - url
        - format
        - cronExpr
        - enabled
        - createdAt
      properties:
        id:
          type: integer
        label:
          type: string
        url:
          type: string
        format:
          type: string
          enum: [mp4, mp3, webm, best]
        quality:
          type: ["string", "null"]
        cronExpr:
          type: string
        enabled:
          type: boolean
        lastRunAt:
          type: ["string", "null"]
        nextRunAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]

    ScheduleInput:
      type: object
      required:
        - label
        - url
        - format
        - cronExpr
      properties:
        label:
          type: string
          minLength: 1
        url:
          type: string
        format:
          type: string
          enum: [mp4, mp3, webm, best]
        quality:
          type: string
        cronExpr:
          type: string
        enabled:
          type: boolean

    ScheduleUpdate:
      type: object
      properties:
        label:
          type: string
        url:
          type: string
        format:
          type: string
          enum: [mp4, mp3, webm, best]
        quality:
          type: string
        cronExpr:
          type: string
        enabled:
          type: boolean

    BatchDownloadInput:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DownloadInput"

    ScheduleRunResponse:
      type: object
      required:
        - dispatched
        - jobs
      properties:
        dispatched:
          type: integer
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Download"