# lib/api-zod/src/generated/types/batchDownloadInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Self-hosted media downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { DownloadInput } from './downloadInput';

export interface BatchDownloadInput {
  items: DownloadInput[];
}