# lib/api-zod/src/generated/types/schedule.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Self-hosted media downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { ScheduleFormat } from './scheduleFormat';

export interface Schedule {
  id: number;
  label: string;
  url: string;
  format: ScheduleFormat;
  /** @nullable */
  quality?: string | null;
  cronExpr: string;
  enabled: boolean;
  /** @nullable */
  lastRunAt?: string | null;
  /** @nullable */
  nextRunAt?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}