# lib/api-zod/src/generated/types/scheduleInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Self-hosted media downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { ScheduleInputFormat } from './scheduleInputFormat';

export interface ScheduleInput {
  /** @minLength 1 */
  label: string;
  url: string;
  format: ScheduleInputFormat;
  quality?: string;
  cronExpr: string;
  enabled?: boolean;
}