# lib/api-zod/src/generated/types/scheduleRunResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Self-hosted media downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { Download } from './download';

export interface ScheduleRunResponse {
  dispatched: number;
  jobs: Download[];
}