# lib/api-zod/src/generated/types/scheduleUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Self-hosted media downloader API
 * OpenAPI spec version: 0.1.0
 */
import type { ScheduleUpdateFormat } from './scheduleUpdateFormat';

export interface ScheduleUpdate {
  label?: string;
  url?: string;
  format?: ScheduleUpdateFormat;
  quality?: string;
  cronExpr?: string;
  enabled?: boolean;
}